allow rich text for plain messages
