+++ extension/js/common/composer/formatters/base-mail-formatter.ts
@@ -1,0 +1,22 @@
+/* © 2016-2018 FlowCrypt Limited. Limitations apply. Contact human@flowcrypcom */
+
+'use strict';
+
+import { NewMsgData } from '../interfaces/composer-types.js';
+import { SendableMsg } from '../../api/email_provider_api.js';
+import { Composer } from '../composer.js';
+
+export interface MailFormatterInterface {
+  sendableMsg(newMsgData: NewMsgData, signingPrv?: OpenPGP.key.Key): Promise<SendableMsg>;
+}
+
+export class BaseMailFormatter {
+
+  protected composer: Composer;
+  protected richText: boolean;
+
+  constructor(composer: Composer) {
+    this.composer = composer;
+    this.richText = composer.sendBtn.popover.choices.richText;
+  }
+}
